Calculator.c

+++ Calculator.c
@@ -79,7 +79,7 @@ int main(int argc, char *argv[])
             // Else Output The Result
             else
             {
-                division = x / y;    // Take x and y Value, Divide Them And Add Result To "result" Value
+                division = (float) x / y;    // Take x and y Value, Divide Them And Add Result To "result" Value
                 printf("Result Is: %.3f\n", division);    // Print Result
                 printf("--------------------------\n");
             }
